requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn
lime
shap
imageio
opencv-python-headless
matplotlib